tutorials/t13_a2a_cloud_run/pyproject.toml
[project]
name = "a2a-tutorial-prod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.3",
    "google-adk>=1.12.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.10",
]
